Rewrite withtout threadpool implementation.
The original implementation attempted to have a single socket, one input
stream, and one output stream serve a client and server connection. This
required the stream to be multiplexed and made thread-safe so that
commands, responses, and file chunks could all be served on the same
stream from multiple threads.

New implementation will have the client create a new NormalConnection
for long-running tasks (GET/PUT commands ending with an `&`). This
    allows each stream to be used for a single thread and message type
    at any given moment.